Widen content_type and retention_days so wire values cannot throw
content_type comes off the wire and varchar(255) throws under strict mode,
which would lose the capture on Phase 4's synchronous insert. retention_days
as tinyint threw on any value over 255 rather than clamping.

+++ database/migrations/2026_08_29_000001_create_endpoints_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->string('name');
             $table->string('token', 64)->unique();
-            $table->unsignedTinyInteger('retention_days')->default(7);
+            $table->unsignedSmallInteger('retention_days')->default(7);
             $table->timestamps();
         });
     }
